requirements.txt
asgiref==3.3.4
Django==3.2.4
djangorestframework==3.12.4
Faker==8.8.1
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
typing-extensions==3.10.0.0
whitenoise==5.2.0
django-cors-headers==3.7.0
gunicorn
